Adapt s390 IUCVTerminal to systemd
IUCVTerminal used to change inittab which doesn't exist anymore.

Instead of inittab the systemd getty templates "serial-getty@.service"
and "iucvtty@.service" are used.
"iucvtty@.service" is provided by s390-tools while
"serial-getty@.service" is provided by systemd.

Currently both templates need to be changed to work with my module but
the change for serial-getty was already accepted upstream so it should
most likely end up in SLES12 and the second is fixing a misleading name
of the iucvtty template for which I have already created a submit
request against s390-tools.

Additionally the terminal emulation selection was removed because it
isn't really needed.

While working on the module I have also replaced serveral unneeded Builtins
and Ops methods but left the rest of the code as it is.