Prevent grey screen in iucvterminal-server module on initial start
If the configuration file /etc/sysconfig/iucv_terminal_server doesn't
exist all tabs were grey which was theoretically fine but it didn't
change besides ZVM/IDs being provided.

So the user had to save the changes first to be able to switch to the
other tabs.

Since there is already a check for ZVM/IDs which prevents a user from
switching to the setting tabs if there are no or faulty ZVM/IDs it is
not needed at all.

So this check was removed.